requirements.txt
openai==0.27.4
git+https://github.com/openai/whisper.git
ffmpeg-python==0.2.0
pyttsx3==2.90
PyAudio==0.2.11
git+https://github.com/openai/whisper.git
boto3==1.26.108
pydub==0.25.1
SpeechRecognition==3.10.0
EdgeGPT==0.1.22.1
torch==1.13.1
scipy==1.10.0

